MosaicDB is the one true resampled image DB
Rather than have StackDB/BlockDB/MosaicDB all be semi-useless subclasses
of ImageLog, lets make just one---MosaicDB---to specialize in holding
any kind of resampled image/mosaic.

One should setup separaete MosaicDB instances/MongoDB collections for
each level of hierarchy, stacks, blocks and mosaics.

As part of this refactoring, I've given MosaicDB the powers of the
FootprintDB: the ability to generate a footprint based on the WCS, and
find overlapping footprints knowing that all images should be resampled
to a common pixel system.

This paves the way for deprecating the FootprintDB.