Use memmap for noise maps of large mosaics
If the variance map surpasses a threshold size, MAX_NPIX_IN_MEMORY,
then memmap is used for astropy.io.fits, and the images are processed
in chunks, with changes flushed to disk after each chunk.

The number of rows in a chunk is set by the MEMMAP_CHUNK_SIZE attribute
in NoiseMapFactory.