Mask pixels in difference images for better offsets
ScalarMosaicFactory.solve_offsets() takes a block_mask_path argument
with BlockDB key for an optional mask image. This is propagated to
Couplings and SliceableImage to mask pixels before differencing.

This lets the user mask objects or image artifacts not covered by the
the weightmap's zeroed pixels (reflection halos are a good case).