Create LockedBlockFactory
This pipeline combines frame stacking and block production into a single
step for cases where we can reasonably assume that all frames in the
image see a uniform sky. This reduces the 'flexibilty' of the block.
(Made for Elixir-LSB images).

Here we demonstrate mosaicing, still need to build in noise mapping.